Test marked Poisson

+++ src/poisson/simulation.jl
@@ -2,7 +2,7 @@ function Base.rand(
     rng::AbstractRNG, pp::AbstractPoissonProcess, tmin::Real=0.0, tmax::Real=1.0
 )
     mark_dist = mark_distribution(pp)
-    N = rand(rng, Poisson(ground_intensity(pp) * (tmax - tmin)))
+    N = rand(rng, Poisson(float(ground_intensity(pp) * (tmax - tmin))))
     times = rand(rng, Uniform(tmin, tmax), N)
     marks = [rand(rng, mark_dist) for n in 1:N]
     return History(; times=times, marks=marks, tmin=tmin, tmax=tmax)
